Move auth -> minerva.auth
