+++ app/src/main/java/be/ugent/zeus/hydra/minerva/auth/MinervaService.java
@@ -1,4 +1,4 @@
-package be.ugent.zeus.hydra.auth;
+package be.ugent.zeus.hydra.minerva.auth;
 
 import android.app.Service;
 import android.content.Intent;
